requirements.txt
torch
requests==2.21.0
tqdm==4.35.0
scipy==1.3.1
fasttext==0.9.1
faiss_cpu==1.6.0
numpy==1.17.2
gensim==3.8.2
faiss==1.5.3
scikit_learn==0.22.2.post1
indic-nlp-library
